listener-upgrade-to-artist-and-permissions
