+++ users/permissions.py
@@ -3,32 +3,33 @@
 from rest_framework.permissions import BasePermission
 
 
-class IsListener(BasePermission):
-    """Пермишен роли слушателя."""
+class ActiveProfilePermission(BasePermission):
+    """Базовый пермишен наличия активного профиля."""
 
+    profile_attr = None
+    message = 'Недостаточно прав.'
+
+    def has_permission(self, request, view):
+        user = request.user
+        if not (user and user.is_authenticated and self.profile_attr):
+            return False
+
+        profile = getattr(user, self.profile_attr, None)
+        return bool(profile and profile.is_active)
+
+
+class IsListener(ActiveProfilePermission):
+    """Пермишен наличия активного профиля слушателя."""
+
+    profile_attr = 'listener_profile'
     message = 'Требуется профиль слушателя.'
 
-    def has_permission(self, request, view):
-        user = request.user
-        return (
-            bool(user and user.is_authenticated)
-            and hasattr(user, 'listener_profile')
-            and user.listener_profile.is_active
-        )
 
+class IsArtist(ActiveProfilePermission):
+    """Пермишен наличия активного профиля артиста."""
 
-class IsArtist(BasePermission):
-    """Пермишен роли артиста."""
-
+    profile_attr = 'artist_profile'
     message = 'Требуется профиль артиста.'
-
-    def has_permission(self, request, view):
-        user = request.user
-        return (
-            bool(user and user.is_authenticated)
-            and hasattr(user, 'artist_profile')
-            and user.artist_profile.is_active
-        )
 
 
 class IsNotArtist(BasePermission):
@@ -50,10 +51,10 @@ class IsStoreObjectOwner(BasePermission):
     message = 'Вы не владелец.'
 
     def has_object_permission(self, request, view, obj):
-        return (
+        return bool(
             request.user
             and request.user.is_authenticated
-            and obj.owner == request.user
+            and obj.owner == request.user,
         )
 
 
@@ -63,8 +64,8 @@ class IsUserObjectOwner(BasePermission):
     message = 'Вы не владелец.'
 
     def has_object_permission(self, request, view, obj):
-        return (
+        return bool(
             request.user
             and request.user.is_authenticated
-            and obj.user == request.user
+            and obj.user == request.user,
         )
