+++ users/schemas/__init__.py
@@ -15,6 +15,7 @@ from .artist_profile import (
     artist_list_schema,
     artist_me_schema,
     artist_public_schema,
+    become_artist_schema,
 )
 from .auth import (
     logout_schema,
@@ -34,6 +35,7 @@ __all__ = [
     'artist_me_schema',
     'artist_public_schema',
     'artist_registration_schema',
+    'become_artist_schema',
     'change_password_schema',
     'change_phone_schema',
     'email_verification_schema',
