+++ users/serializers/__init__.py
@@ -12,11 +12,9 @@ from .artist_profile import (
     ArtistMeSerializer,
     ArtistMeUpdateSerializer,
     ArtistPublicSerializer,
-)
-from .artist_registration import (
-    ArtistRegistrationSerializer,
     BecomeArtistSerializer,
 )
+from .artist_registration import ArtistRegistrationSerializer
 from .base_registration import BaseRegistrationSerializer
 from .listener_profile import ListenerMeSerializer
 from .listener_registration import ListenerRegistrationSerializer
