+++ users/views/__init__.py
@@ -13,6 +13,7 @@ from .artist_profile import (
     ArtistListView,
     ArtistMeView,
     ArtistPublicView,
+    BecomeArtistView,
 )
 from .artist_registration import ArtistRegistrationView
 from .base_registration import BaseRegistrationView
@@ -31,6 +32,7 @@ __all__ = [
     'ArtistMeView',
     'ArtistPublicView',
     'ArtistRegistrationView',
+    'BecomeArtistView',
     'BaseRegistrationView',
     'CustomLogoutView',
     'CustomTokenObtainPairView',
